config_generator/requirements.txt
ollama
pyyaml
